requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
emoji==2.14.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
idna==3.10
jiter==0.7.0
numpy==2.1.2
oauthlib==3.2.2
openai==1.53.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
playwright==1.48.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
